// film-network-ui/src/app/services/models/film-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FilmRequest {
  archive?: boolean;
  director: string;
  filmPoster?: string;
  genre: string;
  id?: number;
  imdbId: string;
  rating?: number;
  synopsis: string;
  title: string;
  year: string;
}